Fixed: 8 byte per memory cell instead of 16

+++ Emulator/src/RAM.hpp
@@ -8,14 +8,14 @@
 class RAM
 {
 private:
-    std::vector<std::uint16_t> m_Memory = std::vector<std::uint16_t>(std::numeric_limits<std::uint16_t>::max() + 1 /* +1 for the 65536 byte */, 0);
+    std::vector<std::uint8_t> m_Memory = std::vector<std::uint8_t>(std::numeric_limits<std::uint16_t>::max() + 1 /* +1 for the 65536 byte */, 0);
 public:
-    inline std::uint16_t GetMemory(std::uint16_t address) const noexcept
+    inline std::uint8_t GetMemory(std::uint16_t address) const noexcept
     {
         return m_Memory[address];
     }
 
-    inline void SetMemory(std::uint16_t address, std::uint16_t value) noexcept
+    inline void SetMemory(std::uint16_t address, std::uint8_t value) noexcept
     {
         m_Memory[address] = value;
     }
